backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
motor==3.7.1
pydantic-settings==2.12.0
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
